% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_compute}
\alias{batch_compute}
\title{Compute batches}
\usage{
batch_compute(n_rows, n_batches)
}
\arguments{
\item{n_rows}{a numeric specifying the number of rows}

\item{n_batches}{a numeric specifying the number of output batches}
}
\value{
a list
}
\description{
Compute batches
}
\examples{

library(textTinyR)

btch = batch_compute(n_rows = 1000, n_batches = 10)
}
