% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bytes_converter}
\alias{bytes_converter}
\title{bytes converter of a text file ( KB, MB or GB )}
\usage{
bytes_converter(input_path_file = NULL, unit = "MB")
}
\arguments{
\item{input_path_file}{a character string specifying the path to the input file}

\item{unit}{a character string specifying the unit. One of \emph{KB}, \emph{MB}, \emph{GB}}
}
\value{
a number
}
\description{
bytes converter of a text file ( KB, MB or GB )
}
\examples{

library(textTinyR)

# bc = bytes_converter(input_path_file = 'some_file.txt', unit = "MB")
}

