% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_distance}
\alias{cosine_distance}
\title{cosine distance of two character strings (each string consists of more than one words)}
\usage{
cosine_distance(sentence1, sentence2, split_separator = " ")
}
\arguments{
\item{sentence1}{a character string consisting of multiple words}

\item{sentence2}{a character string consisting of multiple words}

\item{split_separator}{a character string specifying the delimiter(s) to split the sentence}
}
\value{
a float number
}
\description{
cosine distance of two character strings (each string consists of more than one words)
}
\examples{

library(textTinyR)

sentence1 = 'this is one sentence'

sentence2 = 'this is a similar sentence'

cds = cosine_distance(sentence1, sentence2)
}
