% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Count_Rows}
\alias{Count_Rows}
\title{Number of rows of a file}
\usage{
Count_Rows(PATH, verbose = FALSE)
}
\arguments{
\item{PATH}{a character string specifying the path to a file}

\item{verbose}{either TRUE or FALSE}
}
\value{
a numeric value
}
\description{
Number of rows of a file
}
\details{
This function returns the number of rows for a file. It doesn't load the data in memory.
}
\examples{

library(textTinyR)

PATH = system.file("example_files", "word_vecs.txt", package = "textTinyR")

num_rows = Count_Rows(PATH)
}
