% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dense_2sparse}
\alias{dense_2sparse}
\title{convert a dense matrix to a sparse matrix}
\usage{
dense_2sparse(dense_mat)
}
\arguments{
\item{dense_mat}{a dense matrix}
}
\value{
a sparse matrix
}
\description{
convert a dense matrix to a sparse matrix
}
\examples{

library(textTinyR)

tmp = matrix(sample(0:1, 100, replace = TRUE), 10, 10)

sp_mat = dense_2sparse(tmp)
}
