% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dice_distance}
\alias{dice_distance}
\title{dice similarity of words using n-grams}
\usage{
dice_distance(word1, word2, n_grams = 2)
}
\arguments{
\item{word1}{a character string}

\item{word2}{a character string}

\item{n_grams}{a value specifying the consecutive n-grams of the words}
}
\value{
a float number
}
\description{
dice similarity of words using n-grams
}
\examples{

library(textTinyR)

word1 = 'one_word'

word2 = 'two_words'

dts = dice_distance(word1, word2, n_grams = 2)
}
