% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dims_of_word_vecs}
\alias{dims_of_word_vecs}
\title{dimensions of a word vectors file}
\usage{
dims_of_word_vecs(input_file = NULL, read_delimiter = "\\n")
}
\arguments{
\item{input_file}{a character string specifying a valid path to a text file}

\item{read_delimiter}{a character string specifying the row delimiter of the text file}
}
\value{
a numeric value
}
\description{
dimensions of a word vectors file
}
\details{
This function takes a valid path to a file and a file delimiter as input and estimates the dimensions of the word vectors by using the first row of the file.
}
\examples{

library(textTinyR)

PATH = system.file("example_files", "word_vecs.txt", package = "textTinyR")

dimensions = dims_of_word_vecs(input_file = PATH)
}
