% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{levenshtein_distance}
\alias{levenshtein_distance}
\title{levenshtein distance of two words}
\usage{
levenshtein_distance(word1, word2)
}
\arguments{
\item{word1}{a character string}

\item{word2}{a character string}
}
\value{
a float number
}
\description{
levenshtein distance of two words
}
\examples{

library(textTinyR)

word1 = 'one_word'

word2 = 'two_words'

lvs = levenshtein_distance(word1, word2)
}
