% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sparse_binary}
\alias{load_sparse_binary}
\title{load a sparse matrix in binary format}
\usage{
load_sparse_binary(file_name = "save_sparse.mat")
}
\arguments{
\item{file_name}{a character string specifying the binary file}
}
\value{
loads a sparse matrix from a file
}
\description{
load a sparse matrix in binary format
}
\examples{

\dontrun{

library(textTinyR)

load_sparse_binary(file_name = "save_sparse.mat")

}
}
