% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_sparsity}
\alias{matrix_sparsity}
\title{sparsity percentage of a sparse matrix}
\usage{
matrix_sparsity(sparse_matrix)
}
\arguments{
\item{sparse_matrix}{a sparse matrix}
}
\value{
a numeric value (percentage)
}
\description{
sparsity percentage of a sparse matrix
}
\examples{

library(textTinyR)

tmp = matrix(sample(0:1, 100, replace = TRUE), 10, 10)

sp_mat = dense_2sparse(tmp)

dbl = matrix_sparsity(sp_mat)
}

