% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_characters}
\alias{read_characters}
\title{read a specific number of characters from a text file}
\usage{
read_characters(input_file = NULL, characters = 100, write_2file = "")
}
\arguments{
\item{input_file}{a character string specifying a valid path to a text file}

\item{characters}{a numeric value specifying the number of characters to read}

\item{write_2file}{either an empty string ("") or a character string specifying a valid output file to write the subset of the input file}
}
\description{
read a specific number of characters from a text file
}
\examples{

library(textTinyR)

# txfl = read_characters(input_file = 'input.txt', characters = 100)
}

