% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rows}
\alias{read_rows}
\title{read a specific number of rows from a text file}
\usage{
read_rows(
  input_file = NULL,
  read_delimiter = "\\n",
  rows = 100,
  write_2file = ""
)
}
\arguments{
\item{input_file}{a character string specifying a valid path to a text file}

\item{read_delimiter}{a character string specifying the row delimiter of the text file}

\item{rows}{a numeric value specifying the number of rows to read}

\item{write_2file}{either "" or a character string specifying a valid output file to write the subset of the input file}
}
\description{
read a specific number of rows from a text file
}
\examples{

library(textTinyR)

# txfl = read_rows(input_file = 'input.txt', rows = 100)
}
