% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sparse_binary}
\alias{save_sparse_binary}
\title{save a sparse matrix in binary format}
\usage{
save_sparse_binary(sparse_matrix, file_name = "save_sparse.mat")
}
\arguments{
\item{sparse_matrix}{a sparse matrix}

\item{file_name}{a character string specifying the binary file}
}
\value{
writes the sparse matrix to a file
}
\description{
save a sparse matrix in binary format
}
\examples{

library(textTinyR)

tmp = matrix(sample(0:1, 100, replace = TRUE), 10, 10)

sp_mat = dense_2sparse(tmp)

# save_sparse_binary(sp_mat, file_name = "save_sparse.mat")
}
