% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utf_locale}
\alias{utf_locale}
\title{utf-locale for the available languages}
\usage{
utf_locale(language = "english")
}
\arguments{
\item{language}{a character string specifying the language for which the utf-locale should be returned}
}
\value{
a utf locale
}
\description{
utf-locale for the available languages
}
\details{
This is a limited list of language-locale. The locale depends mostly on the text input.
}
\examples{

library(textTinyR)

utf_locale(language = "english")
}
